pyproject.toml
[project]
name = "lite-star-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksandr Sidorov", email = "a.sidorov@teleguard.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

[tool.poetry]
name = "lite-star-test"
version = "0.1.0"
description = ""
authors = ["Aleksandr Sidorov <a.sidorov@teleguard.com>"]
readme = "README.md"
packages = [{include = "lite_star_test", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.12"
litestar = {extras = ["standard"], version = ">=2.15.2,<3.0.0"}
litestar-granian = "^0.6.0"
litestar-asyncpg = "^0.3.0"
advanced-alchemy = "^0.20.0"
msgspec = "^0.18.6"
python-dotenv = "^1.0.0"
typing-extensions = "^4.12.2"
pydantic = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"